added pytests for store creation
