+++ on_boarding_project/.env
@@ -1,0 +1,3 @@
+
+SECRET_KEY = 6ywy=@ds$n191r=(3c0@x&cm7kt)8lfj9o_5u+k^--_m!wpwis
+SECRET_KEY_1 = "2a1ded95-e016-400c-9988-6d89b60bb366"
