+++ on_boarding_project/on_boarding_project/test_settings.py
@@ -1,0 +1,11 @@
+import os
+from django.contrib.auth import settings
+
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+SECRET_KEY = os.environ['SECRET_KEY']
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
+    }
+}
