+++ on_boarding_project/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = on_boarding_project.settings
+python_files = tests.py test_*.py *_tests.py
